% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trip_funs.R
\name{tripgen}
\alias{tripgen}
\title{Generation time calculation function for triplicate data series}
\usage{
tripgen(csv, OD_col1, OD_col2, OD_col3, midlog)
}
\arguments{
\item{csv}{Filepath to and name of the csv file holding the data.}

\item{OD_col1}{Column holding first column of optical density data.}

\item{OD_col2}{Column holding second column of optical density data.}

\item{OD_col3}{Column holding third column of optical density data.}

\item{midlog}{For the strain of interest, the typical midlog value.}
}
\value{
Generation doubling time in the same unit of time as in the data sheet.
}
\description{
Generation time calculation function for triplicate data series
}
\examples{
tripgen("sample_data_trip.csv", "OD600_1", "OD600_2", "OD600_3", 0.3)
tripgen("sample_data_trip.csv", "OD600_1", "OD600_2", "OD600_3", 0.3)
}
