% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trip_funs.R
\name{tripgraph}
\alias{tripgraph}
\title{Triplicate graphing function.}
\usage{
tripgraph(csv, OD_col1, OD_col2, OD_col3, nm = 600)
}
\arguments{
\item{csv}{Filepath to and name of the csv file holding the data.}

\item{OD_col1}{Column holding first column of optical density data.}

\item{OD_col2}{Column holding second column of optical density data.}

\item{OD_col3}{Column holding third column of optical density data.}

\item{nm}{Wavelength at which optical density was measured; default is 600nm}
}
\description{
Triplicate graphing function.
}
\examples{
tripgraph("sample_data_trip.csv", "OD600_1", "OD600_2", "OD600_3")
tripgraph("sample_data_trip.csv", "OD600_1", "OD600_2", "OD600_3", 200)
}
