% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pilot_funs.R
\name{singen}
\alias{singen}
\title{Generation time calculation function}
\usage{
singen(csv, OD.col, midlog, time.col = "Time")
}
\arguments{
\item{csv}{Filepath to and name of the csv file holding the data.}

\item{OD.col}{Column holding optical density data.}

\item{midlog}{For the strain of interest, the typical midlog value.}

\item{time.col}{Column holding times at which data was collected; default is "Time"}
}
\value{
Generation doubling time in the same unit of time as in the data sheet.
}
\description{
Generation time calculation function
}
\examples{
singen("sample_data.csv", "OD600", 0.3, time.col = "Time")
singen("sample_data.csv", "OD600", 0.4)
}
