% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pilot_funs.R
\name{singraph}
\alias{singraph}
\title{Data graphing function.}
\usage{
singraph(csv, OD_col, nm = 600)
}
\arguments{
\item{csv}{Filepath to and name of the csv file holding the data.}

\item{OD_col}{Column holding optical density data.}

\item{nm}{Wavelength at which optical density was measured; default is 600nm}
}
\value{
Graph with single growth curve.
}
\description{
Data graphing function.
}
\examples{
singraph("sample_data.csv", "OD600")
singraph("sample_data.csv", "OD600", 300)
}
